Test MC Server in GitHub Actions (#425)
I created a fork to add a job to the workflow in which GitHub Actions can runServer to test if it will load a world without crashing. It uses a file called `TESTSERVER.txt` which by default is set to "TEST FAILED". If the server is able to get all the way to the FMLServerStartedEvent, the server changes it to "TEST SUCCESS", after which the server throws an Exception. This behavior is only done when the file `TESTSERVER.txt` is found in the root directory of the Minecraft server.