GitHub Actions will now sign builds (#427)
I forked the repo to test another addition to our GitHub Actions workflow. All builds from this point on will now be signed with our keystore. We will not necessarily have The Midnight check to see if the build is signed, but this will help us ensure that when issues are created for the 1.16.1+ versions of The Midnight, they aren't modified in any way.