Disable the vote screen in the dev environment
I live in the United States, and a lot of mods are asking users to vote for president in our country. Although this is considered to be a noble cause, I do not want to be seeing this screen every time I need to rework the run environment. Shadew did not get this screen because the vote screen grabs your IP and check if you live in the United States.